Test for standard js & use it if available locally
This assumes that standard is not installed system-wide but per
project. This allows it to be a checker on a per project basis.

Note: Sometimes there is a :warning: about an eslint dep & syntastic
doesn't use standard. Usually is fixed by installing the eslint dep but
I'm not sure how I get into that situation.