fix: Remove most dupes in $PATH
The one remaining dupe is the Google Cloud SDK. It runs a script that I
don't think I can paste in place.

Ideas: Plenv is pasted in place and may get out of date because of
this. To test that it's up to date on load, I could capture the
intended eval code and compare it with my current pasted code on zsh
startup. Maybe with a boolen flag for when it's already done, say on
subsequent startups in tmux etc.