fix: Switch to 'fp' as esc mapping in vim
Did some grepping of my dict/words file and found that only 11 words had
that combination of letters. Its unfortunate that its one row up and on
a different hand, but thats much better numbers than the 21,000 words
that had 'ne' in them. Plus having something not on the home row is
expected when using Colemak which is designed to maximize the number of
words you can type using the home row.

I also thought about adding 'pf' as it only had 97 words and might allow
me to forget the direction that I type those two letters. I tried typing
them both ways and found that the pf way was noticeably harder to type.
Why that is I am not sure.