refact: Separate major & minor versions for initial node version
These values should be automatable by looking in the directory for the
largest even major version available. This might even be part of what
the script is already doing as it seems to be searching and sorting in
the versions directory, just with a pattern for the major and minor
version numbers.