fix: Prevent starting zsh w/ path_helper prepended to existing PATH
When starting `zsh` in `tmux`, the PATH was prepended with the
`path_helper` paths. This caused things that were available at the
system level to overwrite version manager paths (like `plenv` &
`rbenv`).

A problem with this fix is that it blows away the PATH even in
sub-shells. I can't think of situations where this would be a problem
but those are famous last words. The assumption I'm making is that the
PATH is curated and somewhat centralized to my .zshrc.