chore: Add bare to nvm lazy cmd & autoload .nvmrc
The nvm plugin kinda conflicts with my custom loading of nvm, but only
causes the nvmrc version to double load once. Afterwards everything
seems to be the nvm omz plugin.

The plugin also prevents the speedup of directly loading the stable
version of node and instead loads up nvm on the first call to node. Not
sure how important this is to me as its not too long and the initial
shell load time is still quick. Might consider removing my custom loader
in the future and maybe keep my autoloading of the .nvmrc so nvm is
loaded automatically in directories with a .nvmrc.